added dividends

+++ src/main/java/org/tradingsignal/pojo/yahoo/Dividend.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class Dividend {
     private double amount;
-    private int date;
+    private long date;
 }
